requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click
colorama==0.4.6
datasets==4.0.0
dill==0.3.8
faiss-cpu==1.12.0
filelock==3.19.1
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2025.3.0
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.1
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx
numpy
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
pyarrow==21.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
pypdf==6.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn
scipy
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
